WIP: Cython
